fault_injection: add event triggering system
This adds an event triggering mechanism.

Basically it will trigger an event with an "event_id" data in a given time and
stop the virtual machine.

A Python script will be able to recognize this fault event and call a given callback.
Finally we will have a notify function in the Python API which will call a Python callback
in a given time which will eg modify the state of the VM (inject GPIO or write memory).

Signed-off-by: KONRAD Frederic <fred.konrad@greensocs.com>
Reviewed-by: Alistair Francis <alistair.francis@xilinx.com>
[FK: 260326 - fixed the documentation to fit 70 columns.
              (a937b6aa739 qapi: Reformat doc comments to conform to current
              conventions)
     260325 - fixed include names
              (32cad1ffb81 include: Rename sysemu/ -> system/)
     260219 - added missing qapi documentation as now required
              (0cec50119f8 qapi: Require member documentation (with loophole))
     - Included some merge fix (dropped qmp-commands.hx)]
Signed-off-by: Frederic Konrad <fkonrad@amd.com>