memattrs: Align with upstream by using requester_id
Upstream calls the master_id field requester_id. Align
with upstream and move all uses of master_id to use
requester_id.

In addition to that, add a QOM property called requester-id.
To avoid breaking existing machine DTBs, we keep the
master-id property but with a warning about future deprecation
when used.

Signed-off-by: Edgar E. Iglesias <edgar.iglesias@xilinx.com>