versal: Introduce support for the Versal RTC
Added the Versal RTC functionality in the same module as the ZU+ RTC.

The version property was introduced to handle the difference in behaviour
betweenthe Zynq Ultrascale+ RTC and the Versal RTC. The property
introduced was:
    version: (string) designates the version of the IP,
             possible values are "1.0.1" and "2.0.0"

Fixed incorrect behaviour for register SET_TIMER. rtc_set_timer_postw()
was introduced to allow the value set to propagate to the tick_offset
count. This mimics closely the expected behaviour of the hardware.

Fixed incorrect behaviour for register  SET_CALB.
rtc_calibration_write_postw() was introduced to allow the value set to
be reflected in the read calibration register. This mimics closely the
behaviour of the hardware.

Set the appropriate value of the reset for the CONTROL register based
on the version of the IP.

Signed-off-by: Sakis Panou <sakis.panou@xilinx.com>
Signed-off-by: Edgar E. Iglesias <edgar.iglesias@xilinx.com>
[FK: 260324 - fixed against property constification and DEFINE_PROP_END_OF_LIST
              removal (5fcabe628b8 include/hw/qdev-properties: Remove
                       DEFINE_PROP_END_OF_LIST)]
Signed-off-by: Frederic Konrad <fkonrad@amd.com>