xlnx-efuse: Add init-factory-tbits property
The property value is default to true.  When true and a tbits
group is in unprogrammed state, the device model sets the bits
to the factory pattern.

The property is used to simulate the Xilinx process of always
pre-programming the T-bits, prior to shipping from factory.

Starting 2019.2 release, Xilinx embeddedsw library will always
treat incorrect T-bits as non-recoverable device error, to catch
device pre-mature failure or mishandling in factory.

Signed-off-by: Tong Ho <tong.ho@xilinx.com>
Signed-off-by: Edgar E. Iglesias <edgar.iglesias@xilinx.com>