hw/ssi/spips: zynqmp: Fix LQSPI cache loading for byte-aligned read
Access to the LQSPI region can be byte-aligned, while the check to see if
we should load a new cache assumed 32-bit alignment.

This meant that if someone were to access a byte-aligned part of memory
close to the end of the cache (for example, 0x3FD), the cache would
refresh with the same memory that was cached before; since the address still
lies on the same page.

Once the cache is refreshed, the transaction is recursively retried with
the "new" cache, meaning the cache will continuously refresh until
we stack overflow.

Signed-off-by: Joe Komlodi <komlodi@xilinx.com>
Message-Id: <1595961525-137936-2-git-send-email-komlodi@xilinx.com>
Signed-off-by: Edgar E. Iglesias <edgar.iglesias@xilinx.com>