hw/ssi/axiqspi: Mark unimplimented commands as such
Some of the commands aren't implemented by the m25p80 model.
Rather than execute them and silently have nothing happen on the flash device,
it's better to say that they're not implemented and not execute them.

This also means that the controller will set the COMMAND_ERROR bit when an
unimplimented command is executed.

Signed-off-by: Joe Komlodi <komlodi@xilinx.com>
Reviewed-by: Francisco Iglesias <francisco.iglesias@xilinx.com>
Message-Id: <1596848872-311038-19-git-send-email-komlodi@xilinx.com>
Signed-off-by: Edgar E. Iglesias <edgar.iglesias@xilinx.com>