smmuv3: Add stage 1 address translation generation
ATOS allows a user to provide a virtual address, which then the SMMU will
return a physical address for.
This is useful when testing to see if a virtual address will map to an
expected physical address.

Since the SMMU only supports stage 1 translation, we can also only support
stage 1 ATOS.

Signed-off-by: Joe Komlodi <joe.komlodi@xilinx.com>
Message-Id: <1626986531-232266-6-git-send-email-joe.komlodi@xilinx.com>
Signed-off-by: Edgar E. Iglesias <edgar.iglesias@xilinx.com>