versal: pmc-global: Add builtin tamper event responder
This extends the pmc-global model with a builtin responder
to device-tampering events, delivered by setting the property
added in this patch.

The property can be set from QMP, HMP, or other devices.

For each tampering event, the builtin responder is bypassed
if there is an input connected to the corresponding tampering
output gpio from the pmc-global device. This use case is also
known as external responder.

Signed-off-by: Tong Ho <tong.ho@amd.com>
[FK: 260701 - added missing qapi/visitor.h header
     260325 - fixed include names
              (32cad1ffb81 include: Rename sysemu/ -> system/)
     260324 - fixed against property constification and DEFINE_PROP_END_OF_LIST
              removal (5fcabe628b8 include/hw/qdev-properties: Remove
                       DEFINE_PROP_END_OF_LIST)]
Signed-off-by: Frederic Konrad <fkonrad@amd.com>