hw/dma/xlnx_csu_dma: change is-dst property default value to false
The legacy csu_stream_dma model has different default value for is-dst.
To keep DTB backward compatibility, align both models to use the legacy
default value.

Explicitly set the is-dst property in zynqmp and versal SoCs to cope
with this change.

This is in preparation for the removal of the legacy csu_stream_dma
model.

Signed-off-by: Luc Michel <luc.michel@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>