hw/dma/xlnx_csu_dma: add support for 128 bits CRC
Sources DMAs keep a CRC value which consist in the sum of fetched
values from the AXI bus. The CRC value width matches the DMA AXI width.

This commit adds support for 128 bits CRCs. It restricts the possible
DMA AXI width to either 4 or 16 (32 bits or 128 bits CRCs respectively).
This matches the real values found in Xilinx SoCs.

When the DMA is 128 bits, four 32 bits CRC registers are exposed instead of
one.

Signed-off-by: Luc Michel <luc.michel@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>