hw: remove OSPI and DMA CSU legacy models
Remove both the xlnx-versal-ospi-v1 and xlnx_csu_dma legacy model.

Add an alias for the DMA model so that DTBs using the legacy
zynqmp.csu-dma name instantiate the new model.

The OSPI model does not need an alias. This commit realigns its name
with upstream (xlnx.versal-ospi) so the legacy name (xlnx,versal-ospi)
will match this new name thanks to "." to "," auto-substitution during
generic FDT machine instantiation.

Note that both are removed at the same time because the interfaces
between the two are not compatible.

Signed-off-by: Luc Michel <luc.michel@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>