hw/mdio: add the transfer method
Add the generic `transfer' method to the MDIO bus. This method
generalizes the existing send and recv methods. It takes an MDIO frame
as parameter. This frame embeds MDIO protocol details and prepares
support for clause 45 extended MDIO.

In case of a read operation, the slave will modify the frame's data
field to return the response to the master.

Signed-off-by: Luc Michel <luc.michel@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>
Reviewed-by: Sai Pavan Boddu <sai.pavan.boddu@amd.com>