hw/mdio: add a basic model of a clause 45 PHY
Add a clause 45 capable PHY device. Clause 45 (extended MDIO) is
required for link speeds >1G. The clause 45 specification allows
multiple devices (MMDs) inside the same "package" (PHY here).

This simple PHY only implements the PMA/PMD device. It advertises speeds
according to the enabled capabilities in the device array. The only
device we expose for now is a non-existant generic clause 45 PHY. This
is enough for Linux to probe it and report available speeds. Note that
auto-negotiation is not implemented.

Signed-off-by: Luc Michel <luc.michel@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>
Reviewed-by: Sai Pavan Boddu <sai.pavan.boddu@amd.com>
[FK: 260324 - fixed class_init() to take a const data argument
              (12d1a768bdf qom: Have class_init() take a const data argument)
     260318 - fixed against class_init prototype change
              (12d1a768bdf qom: Have class_init() take a const data argument)
            - fixed against type_register_static removal
              (281305d3e08 qom/object: Remove type_register())
     260311 - fixed against reset set change
              (e3d0814368d hw: Use device_class_set_legacy_reset() instead of
              opencoding)]
Signed-off-by: Frederic Konrad <fkonrad@amd.com>