target/arm: allow MPIDR feature bits in mp-affinity property
The MPIDR register contains some feature bits that are not part of the
affinity value. At least one of them (the MT bit) is set at the
discretion of the integrator. It indicates that this CPU may be part of a
SMP system including CPUs with multiple threads (it does not imply that
this CPU is actually multi-threaded).

This commit adds support for this bit in the mp-affinity property. The
mp_affinity value is then split into two values, the real mp_affinity (to
not confuse the existing code that relies on plain affinity without
feature bits) and mpidr_feat containing only the supported feature bits.
The two are OR'ed when MPIDR is read.

Signed-off-by: Luc Michel <luc.michel@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>
[FK: 260216 - fixed about kvm64.c to kvm.c merge
              (de3c96017f1 target/arm/kvm: Merge kvm64.c into kvm.c)]
Signed-off-by: Frederic Konrad <fkonrad@amd.com>