hw/misc/xlnx-versal-pmc-global: don't reset persistent storages
Those Persistent Global General Storage registers are only reset during power on
reset.  The reset callback is called at power-on and when the board is warm
reset.  So don't reset those persistent registers in the reset callback,
the behavior will be correct at power on since they will be reset to 0
automagically when the machine is initialized.

Signed-off-by: Frederic Konrad <fkonrad@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>
Reviewed-by: Luc Michel <luc.michel@amd.com>