hw/misc/xlnx-xppu: fix container memory region initialization
The XPPU device has 3 memory regions:
  - One for the memory mapped registers
  - One for the actual memory
  - One container instantiating the two others as sub-regions.

The latter was initialized as I/O MR, with the ops from the memory
mapped registers, and the device state as opaque.

The problem lies in the fact that this container MR has holes in it.
Accessing one of those holes will cause QEMU to crash in the read/write
I/O callback because the expected opaque is not the device state here.

There is no good reason why this container region should be I/O.
Initialize it as a generic MR. An access to a hole will simply be
discarded.

Signed-off-by: Luc Michel <luc.michel@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>
Reviewed-by: Frederic Konrad <fkonrad@amd.com>