hw/misc/xlnx-versal-pmc-iou-slcr: implement the USB power state regs
Implement the CUR_PWR_ST register to match the value of PWR_STATE_REQ.
PWR_STATE_REQ is written to change the power status of the USB
controller. The state is reported in CUR_PWR_ST. Even though we don't
implement the power state, stub the CUR_PWR_ST register to avoid
timeouts.

Note that the PMC IOU SLCR model is duplicated. Do the modification in
both.

Signed-off-by: Luc Michel <luc.michel@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>