xlnx-versal-psm-global: switch apu wakeup as client_gpio
The PSM global is a GPIO controller for the 'wakeup-request'
gpios.  Unfortunately the APU's GIC is also GPIO controller for it's
wake_request signal.  So this switches the apu 'wakeup-request' GPIO from
the controller side to the client side so it can be wired it to the GIC.  Also
the RPU's 'wakeup-request' GPIOs must stay in the controller side since the
intc-redirect is still used for that purpose.  Finally the 'pwr-ctrl' GPIO size
is shrinked from 32 to 6, since the other one are unused at the moment.

Signed-off-by: Frederic Konrad <fkonrad@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>