crypto/ecdsa: add the QCryptoEcdsa abstraction
Add a QCryptoEcdsa API to perform ECDSA operations. For now only the API
is here, and the only implementation is a stub returning
QCRYPTO_ECDSA_ERROR_UNKNOWN for every calls.

Note that this API is device modeling oriented, hence providing
unusual operations such as "recover the public key from the private
key", or "provide the random number k to use for a signing operation".
It also has a quite detailed error reporting for the same reason.

Signed-off-by: Luc Michel <luc.michel@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>