crypto/ecdsa-gcrypt: reimplement the signing op for older libgcrypt
It appears that libgcrypt <1.10 does not support user-supplied k value
during a ECDSA signing operation. This is a problem for us because we
need to provide one, not let gcrypt generate one.

This commit reimplements the ECDSA signing operation by using gcrypt
primitives over elliptic curves. This new implementation is only used
when an incompatible gcrypt version is detected. For newer versions the
gcrypt signing function is still in use.

Note that this implementation is by no mean cryptographically secure.
In particular nothing is done to mitigate timing attacks.

Signed-off-by: Luc Michel <luc.michel@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>