hw/arm/pchannel: add a dummy P-Channel device
Add a dummy device that implements the P-Channel interface. This device
always fulfill state change requests and reports the last state it was
set to.

This device will be useful for single arch PSXC boards where the ARM
CPUs are not present. This device can be used instead to stub the
P-Channel interfaces on the LPD SLCR and the APU PCIL devices.

Signed-off-by: Luc Michel <luc.michel@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>
[FK: 260324 - fixed class_init() to take a const data argument
              (12d1a768bdf qom: Have class_init() take a const data argument)
            - fixed against property constification and DEFINE_PROP_END_OF_LIST
              removal (5fcabe628b8 include/hw/qdev-properties: Remove
                       DEFINE_PROP_END_OF_LIST)
     260311 - fixed against reset set change
              (e3d0814368d hw: Use device_class_set_legacy_reset() instead of
              opencoding)]
Signed-off-by: Frederic Konrad <fkonrad@amd.com>