hw/misc/xlnx-psxc-lpx-slcr: introduce the PSXC LPX SLCR device
Add a skeleton of the PSXC LPX SLCR device. Compared to the PSX
(versal-net) version, it has diverged by a good mean. This comes mainly
from the fact that the PSM does not exist anymore in PSXC. This is why a
new model is introduced for this device instead of tweaking the PSX
version to support both.

Signed-off-by: Luc Michel <luc.michel@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>
[FK: 260324 - fixed class_init() to take a const data argument
              (12d1a768bdf qom: Have class_init() take a const data argument)
     260304 - fixed against reset API changes (ad80e367447 hw, target: Add
              ResetType argument to hold and exit phase methods)]
Signed-off-by: Frederic Konrad <fkonrad@amd.com>