hw/misc/xlnx-psxc-lpx-slcr: implement WAKEUP[0|1] and POWER_DWN IRQs
Implement the IRQs related to WAKEUP0, WAKEUP1 and POWER_DWN registers.
Those IRQs are raised by corresponding P-Channel events on cores and
clusters. For the APUs, the IRQs actually come from the APU_PCIL device
(thus the corresponding apu-[cluster|core]-pchan-[poweroff|wakeup] input
GPIOs). For the RPUs, the RPU_PCIL is embedded into this device and is
not implemented yet. The implementation will come with next commits.

Signed-off-by: Luc Michel <luc.michel@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>