hw/misc/xlnx-psxc-lpx-slcr: implement the RPU_PCIL P-Channel interface
Implement the P-Channel interface in the RPU_PCIL sub-device. This
includes the RPU_PCIL_[A-E][0-1]_PR, _PS and _PA register to communicate
with the interface, and the corresponding IRQ (*_ISR, _IMR, _IEN and
_IDS). A0 and A1 are RPU cluster 0 cores 0 and 1, B0 and B1 cluster 1
and so on.

*_PS is used to set the desired pstate the core should switch to. A
write of 1 to *_PR will effectively trigger the request. The request
result and current pstate can be found in _PA.

The IRQ is raised when the core switches from an OFF to an ON pstate
(from 0 to 2). This IRQ is wired to the WAKEUP1_IRQ_STATUS register.

Signed-off-by: Luc Michel <luc.michel@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>
[FK: 260324 - fixed against property constification and DEFINE_PROP_END_OF_LIST
              removal (5fcabe628b8 include/hw/qdev-properties: Remove
                       DEFINE_PROP_END_OF_LIST)]
Signed-off-by: Frederic Konrad <fkonrad@amd.com>