hw/misc/ipcores-rsa5-4k: cache the r_inv value
The r_inv value is the modular multiplicative invert of r mod m2, used
in Montgomery multiplications. This value is costly to compute. This
commit caches it into the RSA5X state. The cache is invalidated whenever
there is a write into the m2 register memory area, or a related
parameter modification (m2 register position or number of digits).

Signed-off-by: Luc Michel <luc.michel@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>