versal: pmc-sysmon: Fill info only for enabled sensors
This patch ensures that config info for a sensor is filled only if:
1. The satellite itself is running, and
2. The sensor itself is enabled.

For any voltage sensor other than the VCCINT one, it is enabled
only if it is included in the ADC-sampling list.  In other words,
such a sensor will not report measurements to ams-root unless
it is in the list.

Normally, all these configs are established by a design's cdo.bin,
which is generated by Vivado upon the design's instantiation.

With this change, when running any guest software depending on
sysmon's operation, the respective cdo.bin must be loaded
(typically by PLM). Failure to load will most likely result
in guest software hanging.

Signed-off-by: Tong Ho <tong.ho@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>