hw/crypto/xlnx-asu-kv: implement key clearing
Implement the key clearing operation and status in the ASU keyvault
model. This includes the ASU_KEY_CLEAR register used to clear one or
more keys, and the KEY_ZEROED_STATUS that reports the zeroed keys.

Note that the AES_KEY_ZEROIZE/ZEROED fields are not implemented yet as
they refer to the ASU AES zeroization operation.

Signed-off-by: Luc Michel <luc.michel@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>